requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
fastmcp
fastapi
uvicorn
sqlalchemy
gradio
httpx
Faker
google-genai